monitoring/requirements.txt
pandas
scikit-learn
joblib
streamlit
datetime
matplotlib.pyplot
json
os